ExtractStreamerInfo
Solved issues with extracting streamer infos
Now working on getting the "CreateMember" fucntion implemented to build up the streamers stroed in the streamer info. Need to implement several more methods in TBuffer.  Also there is an outstanding issue with the UnZip in some cases - not sure yet what is happening